Update package & coordinate to be develocity
