--- src/main/java/com/gradle/enterprise/conventions/customvalueprovider/ScanCustomValueNames.java
@@ -1,10 +1,0 @@
-package com.gradle.enterprise.conventions.customvalueprovider;
-
-public interface ScanCustomValueNames {
-    String GIT_COMMIT_NAME = "gitCommitId";
-    String GIT_STATUS = "gitStatus";
-    String GIT_BRANCH_NAME = "gitBranchName";
-    String BUILD_ID = "buildId";
-    String IDEA_VERSION = "ideaVersion";
-    String WATCH_FILE_SYSTEM = "watchFileSystem";
-}
