+++ src/main/java/io/github/gradle/conventions/customvalueprovider/BuildCacheCustomValueProvider.java
@@ -1,0 +1,17 @@
+package io.github.gradle.conventions.customvalueprovider;
+
+import com.gradle.develocity.agent.gradle.scan.BuildScanConfiguration;
+import org.gradle.api.initialization.Settings;
+
+public class BuildCacheCustomValueProvider extends BuildScanCustomValueProvider {
+    public BuildCacheCustomValueProvider(DevelocityConventions conventions) {
+        super(conventions);
+    }
+
+    @Override
+    public void accept(Settings settings, BuildScanConfiguration buildScan) {
+        if (settings.getGradle().getStartParameter().isBuildCacheEnabled()) {
+            buildScan.tag("CACHED");
+        }
+    }
+}
