+++ src/main/java/io/github/gradle/conventions/customvalueprovider/CITagProvider.java
@@ -1,0 +1,20 @@
+package io.github.gradle.conventions.customvalueprovider;
+
+import com.gradle.develocity.agent.gradle.scan.BuildScanConfiguration;
+import org.gradle.api.initialization.Settings;
+
+public class CITagProvider extends BuildScanCustomValueProvider {
+    public CITagProvider(DevelocityConventions conventions) {
+        super(conventions);
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return getConventions().isCiServer();
+    }
+
+    @Override
+    public void accept(Settings settings, BuildScanConfiguration buildScan) {
+        buildScan.tag("CI");
+    }
+}
