+++ src/main/java/io/github/gradle/conventions/customvalueprovider/ScanCustomValueNames.java
@@ -1,0 +1,10 @@
+package io.github.gradle.conventions.customvalueprovider;
+
+public interface ScanCustomValueNames {
+    String GIT_COMMIT_NAME = "gitCommitId";
+    String GIT_STATUS = "gitStatus";
+    String GIT_BRANCH_NAME = "gitBranchName";
+    String BUILD_ID = "buildId";
+    String IDEA_VERSION = "ideaVersion";
+    String WATCH_FILE_SYSTEM = "watchFileSystem";
+}
