--- src/test/java/com/gradle/enterprise/conventions/customvalueprovider/WatchFilesystemCustomValueProviderIntegrationTest.java
@@ -1,20 +1,0 @@
-package com.gradle.enterprise.conventions.customvalueprovider;
-
-import com.gradle.enterprise.fixtures.AbstractDevelocityPluginIntegrationTest;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-public class WatchFilesystemCustomValueProviderIntegrationTest extends AbstractDevelocityPluginIntegrationTest {
-    @ParameterizedTest
-    @CsvSource({
-        "true, ENABLED",
-        "false, DISABLED"
-    })
-    public void addWatchFsCustomValue(Boolean watchFsEnabled, String status) {
-        succeeds("help", watchFsEnabled ? "--watch-fs" : "--no-watch-fs");
-
-        assertTrue(getConfiguredBuildScan().containsValue("watchFileSystem", status));
-    }
-}
