+++ src/test/java/io/github/gradle/conventions/customvalueprovider/GitInformationCustomValueProviderIntegrationTest.java
@@ -1,0 +1,48 @@
+package io.github.gradle.conventions.customvalueprovider;
+
+import io.github.gradle.fixtures.AbstractDevelocityPluginIntegrationTest;
+import org.junit.jupiter.api.Test;
+
+import static io.github.gradle.conventions.customvalueprovider.DevelocityConventions.execAndGetStdout;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class GitInformationCustomValueProviderIntegrationTest extends AbstractDevelocityPluginIntegrationTest {
+    @Test
+    public void doNothingIfNotAGitRepo() {
+        succeeds("help");
+
+        assertFalse(getConfiguredBuildScan().containsBackgroundTag("dirty"));
+        assertFalse(getConfiguredBuildScan().containsBackgroundValue("gitStatus"));
+        assertFalse(getConfiguredBuildScan().containsBackgroundValue("gitBranchName"));
+    }
+
+    @Test
+    public void tagDirtyIfGitRepoIsDirty() {
+        execAndGetStdout(projectDir, "git", "init");
+        execAndGetStdout(projectDir, "git", "add", ".");
+
+        succeeds("help");
+
+        assertTrue(getConfiguredBuildScan().containsBackgroundTag("dirty"));
+        assertTrue(getConfiguredBuildScan().containsBackgroundValue("gitStatus"));
+        assertFalse(getConfiguredBuildScan().containsBackgroundValue("gitBranchName"));
+    }
+
+    @Test
+    public void addGitBranchNameIfAvailable() {
+        write(".gitignore", "*", "!fileToCommit.txt");
+        write("fileToCommit.txt", "hello");
+        execAndGetStdout(projectDir, "git", "init");
+        execAndGetStdout(projectDir, "git", "config", "--add", "remote.origin.url", "https://github.com/gradle/gradle.git");
+        execAndGetStdout(projectDir, "git", "checkout", "-b", "new-branch");
+        execAndGetStdout(projectDir, "git", "add", ".");
+        execAndGetStdout(projectDir, "git", "commit", "-m", "Initial commit");
+
+        succeeds("help");
+
+        assertFalse(getConfiguredBuildScan().containsBackgroundTag("dirty"));
+        assertFalse(getConfiguredBuildScan().containsBackgroundValue("gitStatus"));
+        assertTrue(getConfiguredBuildScan().containsBackgroundValue("gitBranchName", "new-branch"));
+    }
+}
