+++ src/test/resources/META-INF/gradle-plugins/com.gradle.develocity.properties
